refactor(wallet): Split wallet file up PE-608
