+++ src/ardrive.test.ts
@@ -17,7 +17,7 @@ import { ByteCount, UnixTime, stubTransactionID, W, FeeMultiple, TipType } from 
 import { readJWKFile } from './utils/common';
 import { stubEntityID } from './utils/stubs';
 import { expectAsyncErrorThrow } from './utils/test_helpers';
-import { WalletDAO } from './wallet';
+import { WalletDAO } from './wallet_dao';
 
 describe('ArDrive class', () => {
 	let arDrive: ArDrive;
